requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.16.0
anyio==4.2.0
arxiv==2.1.0
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dataclasses-json==0.6.4
distro==1.9.0
ecdsa==0.18.0
extra-streamlit-components==0.1.60
fastapi==0.109.2
feedparser==6.0.10
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.7
langchain-cli==0.0.21
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langchainhub==0.1.14
langserve==0.0.41
langsmith==0.0.87
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.9.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.14
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pathspec==0.12.1
pillow==10.2.0
pinecone-client==3.0.3
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sse-starlette==1.8.2
starlette==0.36.3
streamlit==1.31.1
streamlit-authenticator==0.3.1
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.0
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.23.2
validators==0.22.0
whisper==1.1.10
wikipedia==1.4.0
yarl==1.9.4
zipp==3.17.0